==> projects/m5/01_brief_eda.ipynb <==
---
output-file: brief_eda.html
title: Brief EDA

---



import os
import pandas as pd

# For fast testing for Continuous Integration
PATH_DATA = 'data'
PATH_DATA_RAW = 'data/raw'
os.listdir(PATH_DATA_RAW)

# # no_test
# PATH_DATA = 'data'
# PATH_DATA_RAW = 'data/raw'
# os.listdir(PATH_DATA_RAW)

chunks = pd.read_csv(os.path.join(PATH_DATA_RAW, 'sales_train_evaluation.csv'), chunksize=1000)
df_stv = pd.concat(list(chunks)) # Safe for low RAM situation
df_cal = pd.read_csv(os.path.join(PATH_DATA_RAW, 'calendar.csv'))
df_prices = pd.read_csv(os.path.join(PATH_DATA_RAW, 'sell_prices.csv'))
df_ss = pd.read_csv(os.path.join(PATH_DATA_RAW, 'sample_submission.csv'))

**sales_train_evaluation**:
We have six hierarchical categores and 1941 days of sales data for each of the 30490 items

display(df_stv.head())
display(df_stv.info())

**calendar** gives us some date information. This includes special events like holidays and information about SNAP (ebt foodstamps) acceptance for each day of the training and evaluation periods. 

display(df_cal.head())
display(df_cal.info())

**sell_prices** gives the weekly price for each store/item combination.

display(df_prices.head())
display(df_prices.info())

**sample_submission** shows the expected format of a submission. Each item/store combination will have a row with the 28 day forecast as the columns. We submit predictions for the validation period and the test period in the same file. 

display(df_ss.head())
display(df_ss.info())

## Saving small data
To utilize continuous integration on github, and also use nbdev to test all of our code, we will create a truncated version of our files. They need to be less than 100MB to fit on github and allow fast tests.

Lets only save the last 28 * 6 days of sales data for one item from each dept_id / store_id combination. 

#| no_test
df_len = df_stv.shape[1]
df_stv_small = df_stv.iloc[:, list(range(6)) + list(range(df_len - 28 * 5, df_len))]
original_col_order = df_stv_small.columns
df_stv_small = df_stv_small.groupby(['dept_id', 'store_id']).first().reset_index()[original_col_order]
print(f'Now there are only {df_stv_small.shape[0]} rows and {df_stv_small.shape[1]} columns')
df_stv_small.head(3)

For the small calendar file, we only need data for days greater than or equal to 1802

#| no_test
min_wm_yr_wk = df_cal[df_cal.d == 'd_1802'].wm_yr_wk.values[0]
df_cal_small = df_cal[df_cal.wm_yr_wk >= min_wm_yr_wk]
df_cal_small.info()

For the small prices file, we can filter using the items we have in df_stv_small, and the minimun wm_yr_wk we just made.

#| no_test
item_mask = df_prices.item_id.isin(set(df_stv_small.item_id.tolist()))
date_mask = df_prices.wm_yr_wk >= min_wm_yr_wk
df_prices_small = df_prices[item_mask & date_mask]
df_prices_small.info()

The sample submission should only have the items in df_stv_small

#| no_test
item_list = df_stv_small.id.to_list() + df_stv_small.id.str.replace('evaluation', 'validation').to_list()
df_ss_small = df_ss[df_ss.id.isin(item_list)]

#| no_test
os.makedirs('small_data/raw', exist_ok=True)
df_stv_small.to_csv(os.path.join('small_data/raw', 'sales_train_evaluation.csv'), index=False)
df_cal_small.to_csv(os.path.join('small_data/raw', 'calendar.csv'), index=False)
df_prices_small.to_csv(os.path.join('small_data/raw', 'sell_prices.csv'), index=False)
df_ss_small.to_csv(os.path.join('small_data/raw', 'sample_submission.csv'), index=False)
